backend/requirements.txt
Flask==3.0.0
web3==6.15.1
pycryptodome==3.19.1
flask-cors==4.0.0
python-dotenv==1.0.0
